Izmena lekara

+++ SpringWebApp/src/main/java/isamrs/repository/SalaRepository.java
@@ -1,21 +1,18 @@
 package isamrs.repository;
 
-import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Optional;
 
-import org.springframework.data.domain.Example;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
+
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.Query;
 
+import isamrs.domain.Poseta;
+import isamrs.domain.Pregled;
 import isamrs.domain.Sala;
 
 
-public interface SalaRepository extends JpaRepository<Sala, Long>{
+public interface SalaRepository extends JpaRepository<Sala, Integer>{
+	
 	
 	
 }
